refactor: moved code around after merging base
